+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added the state registration number to cost center schema
+
 ## [0.23.3] - 2022-10-17
 
 ### Changed